requirements.txt
Flask==3.0.3
flask-cors==5.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
httpx==0.27.0

langchain-groq==0.1.9
langchain-core==0.2.28
langchain-community==0.2.10
groq==0.4.1